Role classification and file reorganisation
