+++ X-Server.XPlane/Sources/XPlaneManagers/DatarefManager.cpp
