+++ X-Server.XPlane/X-Server.XPlane.vcxproj
@@ -171,7 +171,7 @@
   <ItemGroup>
     <ClCompile Include="Sources\Callbacks.cpp" />
     <ClCompile Include="Sources\Datarefs\Dataref.cpp" />
-    <ClCompile Include="Sources\Datarefs\DatarefManager.cpp" />
+    <ClCompile Include="Sources\XPlaneManagers\DatarefManager.cpp" />
     <ClCompile Include="Sources\Datarefs\FFA320Dataref.cpp" />
     <ClCompile Include="Sources\main.cpp" />
     <ClCompile Include="Sources\MasterCallback.cpp" />
@@ -181,7 +181,7 @@
     <ClInclude Include="Sources\Callbacks.h" />
     <ClInclude Include="Sources\Datarefs\AbstractDataref.h" />
     <ClInclude Include="Sources\Datarefs\Dataref.h" />
-    <ClInclude Include="Sources\Datarefs\DatarefManager.h" />
+    <ClInclude Include="Sources\XPlaneManagers\DatarefManager.h" />
     <ClInclude Include="Sources\Datarefs\FFA320Dataref.h" />
     <ClInclude Include="Sources\MasterCallback.h" />
     <ClInclude Include="Sources\Tools\SharedValue.h" />
